+++ Panels/tests/WizardPanel-specs.js
@@ -23,6 +23,27 @@ describe('WizardPanel Specs', () => {
 	);
 
 	test(
+		'should have subtitle overridden by subtitle set in `View`',
+		() => {
+			const wizardSubtitle = 'WizardPanel subtitle';
+			const viewSubtitle = 'View subtitle';
+
+			const wizardPanel = mount(
+				<WizardPanel subtitle={wizardSubtitle}>
+					<View subtitle={viewSubtitle} />
+				</WizardPanel>
+			);
+
+			const headerSubtitle = wizardPanel.find('Header').prop('subtitle');
+
+			const expected = viewSubtitle;
+			const actual = headerSubtitle;
+
+			expect(actual).toBe(expected);
+		}
+	);
+
+	test(
 		'should have title overridden by title set in `View`',
 		() => {
 			const wizardTitle = 'WizardPanel title';